% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{cook_example}
\alias{cook_example}
\title{Example data: partial rankings}
\format{
A data frame of 20 rows and 3 columns
\describe{
\item{Item}{Character values giving one of 6 items}
\item{Reviewer}{Character values giving one of 5 reviewers}
\item{Ranking}{Numeric values giving a rank position}
}
}
\source{
Cook, W.D., Golany, B., Penn, M. and Raviv, T., 2007.
Creating a consensus ranking of proposals from reviewers’
partial ordinal rankings. Computers & Operations Research,
34, pp.954-965.
}
\usage{
cook_example
}
\description{
A dataset containing 5 partial rankings
of 6 items. This is the example used by
Cook et al (2007).
}
\keyword{datasets}
