% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{upperBound}
\alias{upperBound}
\title{Upper bound value}
\usage{
upperBound(ccr, umat)
}
\arguments{
\item{ccr}{a vector of rank positions that is a candidate complete ranking}

\item{umat}{a \code{matrix}, which is either the output of \code{\link{evaluationMatrix}},
or a subset of the output of \code{\link{evaluationMatrix}}.}
}
\value{
A numeric value  for the upper bound of a candidate complete ranking
}
\description{
This function is called by \code{\link{RankAggregator}}::\code{\link{consensusRanking}}.
The upper bound value is the value used by the branch and bound
algorithm in determining whether or not to replace the current incumbent
solution.
}
